# openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "LeadRadar API",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/leads": {
      "get": {
        "tags": [
          "leads"
        ],
        "summary": "List Leads",
        "operationId": "list_leads_api_v1_leads_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "service_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Service Id"
            }
          },
          {
            "name": "tier",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tier"
            }
          },
          {
            "name": "country",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Country"
            }
          },
          {
            "name": "industry",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Industry"
            }
          },
          {
            "name": "min_priority",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "number"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Min Priority"
            }
          },
          {
            "name": "has_new",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Has New"
            }
          },
          {
            "name": "q",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "name": "sort",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "default": "priority:desc",
              "title": "Sort"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "default": 1,
              "title": "Page"
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 100,
              "minimum": 1,
              "default": 20,
              "title": "Page Size"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PaginatedResponse_LeadListItem_"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/leads/export.csv": {
      "get": {
        "tags": [
          "leads"
        ],
        "summary": "Export Leads Csv",
        "operationId": "export_leads_csv_api_v1_leads_export_csv_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "service_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Service Id"
            }
          },
          {
            "name": "tier",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tier"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/leads/{company_id}": {
      "get": {
        "tags": [
          "leads"
        ],
        "summary": "Get Lead Detail",
        "operationId": "get_lead_detail_api_v1_leads__company_id__get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "company_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Company Id"
            }
          },
          {
            "name": "service_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string",
                  "format": "uuid"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Service Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LeadDetail"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/leads/{company_id}/outreach": {
      "post": {
        "tags": [
          "leads"
        ],
        "summary": "Generate Lead Outreach",
        "operationId": "generate_lead_outreach_api_v1_leads__company_id__outreach_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "company_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Company Id"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/OutreachGenerateIn"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OutreachDraftOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LeadDetail": {
        "properties": {
          "company": {
            "$ref": "#/components/schemas/CompanyOut"
          },
          "service": {
            "additionalProperties": true,
            "type": "object",
            "title": "Service"
          },
          "score": {
            "additionalProperties": true,
            "type": "object",
            "title": "Score"
          },
          "signals_by_question": {
            "items": {
              "$ref": "#/components/schemas/QuestionSignals"
            },
            "type": "array",
            "title": "Signals By Question",
            "default": []
          },
          "decision_makers": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Decision Makers",
            "default": []
          },
          "history": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "History",
            "default": []
          },
          "sources_summary": {
            "additionalProperties": {
              "type": "integer"
            },
            "type": "object",
            "title": "Sources Summary",
            "default": {}
          }
        },
        "type": "object",
        "required": [
          "company",
          "service",
          "score"
        ],
        "title": "LeadDetail"
      },
      "LeadListItem": {
        "properties": {
          "company": {
            "$ref": "#/components/schemas/CompanyOut"
          },
          "service_id": {
            "type": "string",
            "format": "uuid",
            "title": "Service Id"
          },
          "score": {
            "$ref": "#/components/schemas/ScoreSummary"
          },
          "top_reasons": {
            "items": {
              "additionalProperties": true,
              "type": "object"
            },
            "type": "array",
            "title": "Top Reasons",
            "default": []
          },
          "flags": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Flags",
            "default": []
          },
          "signals_count": {
            "type": "integer",
            "title": "Signals Count",
            "default": 0
          },
          "new_signals_7d": {
            "type": "integer",
            "title": "New Signals 7D",
            "default": 0
          },
          "last_signal_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Signal At"
          },
          "analyzed_at": {
            "anyOf": [
              {
                "type": "string",
                "format": "date-time"
              },
              {
                "type": "null"
              }
            ],
            "title": "Analyzed At"
          }
        },
        "type": "object",
        "required": [
          "company",
          "service_id",
          "score"
        ],
        "title": "LeadListItem"
      },
...
      "OutreachDraftOut": {
        "properties": {
          "channel": {
            "type": "string",
            "title": "Channel"
          },
          "subject": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Subject"
          },
          "body": {
            "type": "string",
            "title": "Body"
          },
          "referenced_signals": {
            "items": {
              "type": "string",
              "format": "uuid"
            },
            "type": "array",
            "title": "Referenced Signals",
            "default": []
          },
          "referenced_quotes": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Referenced Quotes",
            "default": []
          },
          "hook": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Hook"
          },
          "call_to_action": {
            "type": "string",
            "title": "Call To Action"
          },
          "language": {
            "type": "string",
            "title": "Language",
            "default": "en"
          }
        },
        "type": "object",
        "required": [
          "channel",
          "body",
          "call_to_action"
        ],
        "title": "OutreachDraftOut"
      },
      "OutreachGenerateIn": {
        "properties": {
          "service_id": {
            "anyOf": [
              {
                "type": "string",
                "format": "uuid"
              },
              {
                "type": "null"
              }
            ],
            "title": "Service Id"
          },
          "channel": {
            "type": "string",
            "title": "Channel",
            "default": "email"
          },
          "language": {
            "type": "string",
            "title": "Language",
            "default": "en"
          },
          "tone": {
            "type": "string",
            "title": "Tone",
            "default": "professional"
          },
          "sender_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sender Name"
          },
          "sender_title": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sender Title"
          },
          "sender_company": {
            "type": "string",
            "title": "Sender Company",
            "default": "LeadRadar"
          }
        },
        "type": "object",
        "title": "OutreachGenerateIn"
      },
...
      "PaginatedResponse_LeadListItem_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/LeadListItem"
            },
            "type": "array",
            "title": "Items"
          },
          "total": {
            "type": "integer",
            "title": "Total"
          },
          "page": {
            "type": "integer",
            "title": "Page"
          },
          "page_size": {
            "type": "integer",
            "title": "Page Size"
          }
        },
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "page_size"
        ],
        "title": "PaginatedResponse[LeadListItem]"
      },
...
      "QuestionSignals": {
        "properties": {
          "question": {
            "additionalProperties": true,
            "type": "object",
            "title": "Question"
          },
          "strength": {
            "type": "number",
            "title": "Strength"
          },
          "points": {
            "type": "number",
            "title": "Points"
          },
          "signals": {
            "items": {
              "$ref": "#/components/schemas/SignalItem"
            },
            "type": "array",
            "title": "Signals",
            "default": []
          }
        },
        "type": "object",
        "required": [
          "question",
          "strength",
          "points"
        ],
        "title": "QuestionSignals"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}